Rename DatabaseMigrator to DatabaseSeeding and update contract permissions
- Rename the DatabaseMigrator project, namespace, and associated files to DatabaseSeeding to more accurately reflect its role in the solution.
- Update `0056_AdditionalPermissions.sql` to restrict 'Contract Data' permissions to the 'System Support' role only, removing 'SMT' and 'CFO QA Manager' roles.
- Add `aspire.config.json` and bump `CommunityToolkit.Aspire.Hosting.SqlDatabaseProjects` to version 13.2.1-beta.532.
- Update Cake build scripts and Aspire AppHost references to reflect the project renaming.
