+++ Directory.Packages.props
@@ -25,7 +25,7 @@
     <PackageVersion Include="BlazorTime" Version="1.0.3" />
     <PackageVersion Include="ClosedXML" Version="0.105.0" />
     <PackageVersion Include="CodeBeam.MudBlazor.Extensions" Version="9.0.4" />
-    <PackageVersion Include="CommunityToolkit.Aspire.Hosting.SqlDatabaseProjects" Version="13.1.1" />
+    <PackageVersion Include="CommunityToolkit.Aspire.Hosting.SqlDatabaseProjects" Version="13.2.1-beta.532" />
     <PackageVersion Include="Dapper" Version="2.1.72" />
     <PackageVersion Include="FluentValidation" Version="12.1.1" />
     <PackageVersion Include="FluentValidation.DependencyInjectionExtensions" Version="12.1.1" />
